requirements.txt
flask
flask_pymongo
awsgi